Move production code out into own file.
