+++ test/scripts/copy-and-minify-json-files-spec.js
@@ -4,22 +4,9 @@ import {
   promiseThat, is, fulfilled, rejected,
   allOf, truthy
 } from 'hamjest';
-
+import {convertOneFile} from '../../scripts/minify-json-files';
 import fs from 'fs';
 import path from 'path';
-import promisify from 'es6-promisify';
-
-function convertOneFile(fileName, destFileName) {
-  const fileReadPromise = promisify(fs.readFile);
-  const fileWritePromise = promisify(fs.writeFile);
-  return fileReadPromise(fileName).then(fileContent => 
-    fileWritePromise(destFileName, minifyJson(fileContent))
-  );
-}
-
-function minifyJson(content) {
-  return JSON.stringify(JSON.parse(content));
-}
 
 describe('converting', () => {
